+++ client/src/game/ui/HanabiCard.ts
@@ -25,7 +25,7 @@ import * as HanabiCardInit from "./HanabiCardInit";
 import * as HanabiCardMouse from "./HanabiCardMouse";
 import { animate } from "./konvaHelpers";
 import LayoutChild from "./LayoutChild";
-import * as noteIdentity from "./noteIdentity";
+import checkNoteImpossibility from "./noteCheckImpossibility.ts";
 import * as notes from "./notes";
 
 const DECK_BACK_IMAGE = "deck-back";
@@ -1112,7 +1112,8 @@ export default class HanabiCard extends Konva.Group implements NodeWithTooltip {
   }
 
   checkSpecialNote(): void {
-    noteIdentity.checkNoteImpossibility(this.variant, this.state, this.note);
+    const note = globals.state.notes.ourNotes[this.state.order];
+    checkNoteImpossibility(this.variant, this.state, note);
 
     // Morph the card if it has an "exact" card note
     // (or clear the bare image if the note was deleted/changed)
